% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{comparison_data}
\alias{comparison_data}
\title{Comparison data}
\format{
An object of class \code{list} of length 13.
}
\usage{
comparison_data
}
\description{
A list of data frames, each describing data that can be used in comparisons.
See \verb{attr(*, "meta")$Source} of each data frame for its source.
}
\examples{
# Metadata for one of the data frames
print( attr(comparison_data[[1]], "meta") )

# Summary of all available data
ret = NULL
for (nn in names(comparison_data)) {
  r1 = data.frame(Label = nn
    , About = strtrim( attr(comparison_data[[nn]], "meta")$About, 25)
    , `Group 1` = comparison_data[[nn]]$group[1]
    , `Group 2` = comparison_data[[nn]]$group[2]
    , Source = strtrim( attr(comparison_data[[nn]], "meta")$Source, 25)
    , check.names = FALSE)
  ret = rbind(ret, r1)
}
print(ret)
}
\keyword{datasets}
