% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_bin.R
\name{d_bin}
\alias{d_bin}
\title{Dominance effect sizes for binary variables}
\usage{
d_bin(x, meta = NULL)
}
\arguments{
\item{x}{data. See details.}

\item{meta}{either \code{NULL} or a named list.}
}
\value{
Estimates with verbose explanations.
}
\description{
Calculate dominance-based effect sizes for binary variables.
}
\details{
Calculate the following:
\itemize{
\item risk difference, which, for binary variables, is a special case of stochastic difference (aka Glass rank-biserial correlation);
\item adjusted stochastic difference (aka Goodman and Kruskal's gamma);
\item number needed to treat;
\item dominance (aka common language effect size);
\item measure of stochastic superiority (aka Vargha and Delaney's A);
\item adjusted dominance;
\item odds ratio, which, for binary variables, is a special case of generalized odds ratio; and
\item win odds.
}

\code{x} can contain either known parameter values, summary statistics, or microdata.
\itemize{
\item If \code{x} contains known parameter values or summary statistics, it should be
a \code{data.frame} with 2 rows. It can optionally contain a variable called \code{group},
indicating the name of the group. In addition, it must contain one of the following:
\itemize{
\item event probability (\code{p}), which is a known parameter; or
\item the number of events (\code{k}) (also called "successes") and the number of experiments or trials (\code{n}); or
\item proportion of events (\code{prop}) and \code{n}; or
\item \code{prop} and its standard error (\code{se}).
}
\item If \code{x} contains microdata, it should be a named \code{list} with 2 elements, each
being a vector of \code{logical} data.
}

Optionally, \code{x} can have an attribute called \code{"meta"}, which must be a named list.
}
\examples{
# Number of events and number of trials
d_bin(comparison_data$UTI)
d_bin(comparison_data$DVT)
d_bin(comparison_data$MI)

# Proportion estimates and their standard errors (SE)
d_bin(comparison_data$drug.use)

# Known parameter values (based on the UTI example)
d_bin(data.frame(p = c(11/13, 2/8)))
}
\seealso{
Other es: 
\code{\link{d_cont}()},
\code{\link{d_np}()},
\code{\link{d_one}()},
\code{\link{d_ord}()},
\code{\link{d_pair}()}
}
\concept{es}
