% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_cont.R
\name{d_cont}
\alias{d_cont}
\title{Dominance effect sizes for continuous variables}
\usage{
d_cont(x, meta = NULL)
}
\arguments{
\item{x}{data. See details.}

\item{meta}{either \code{NULL} or a named list.}
}
\value{
Estimates with verbose explanations.
}
\description{
Calculate dominance-based effect sizes for continuous (normally distributed)
variables.
}
\details{
Calculate the following:
\itemize{
\item dominance (aka common language effect size);
\item stochastic difference (aka Glass rank-biserial correlation);
\item odds, which, for continuous variables, is a special case of generalized odds ratio; and
\item standardized mean difference (aka Cohen's d).
}

For continuous variables, the following effect sizes simplify to one of the
above-mentioned effect sizes. Thus, they are not reported separately.
\itemize{
\item measure of stochastic superiority (aka Vargha and Delaney's A), which simplifies to dominance;
\item adjusted dominance, which simplifies to dominance;
\item adjusted stochastic difference (aka Goodman and Kruskal's gamma), which simplifies to stochastic difference; and
\item win odds, which simplifies to odds.
}

\code{x} can contain either known parameter values, summary statistics, or microdata.
\itemize{
\item If \code{x} contains known parameter values or summary statistics, it should be
a \code{data.frame} with 2 rows. It can optionally contain a variable called \code{group},
indicating the name of the group. In addition, it must contain one of the following:
\itemize{
\item expected value (\code{mu}) and standard deviation (\code{sigma}), which are known parameters; or
\item sample mean (\code{xbar}), sample standard deviation (\code{sd}), and sample size (\code{n}); or
\item \code{xbar}, standard error of the mean (\code{sem}), and \code{n}.
}

\code{sd} is the "sample" standard deviation, with denominator \code{n - 1}. This is
what is returned by the \code{sd()} function.
\item If \code{x} contains microdata, it should be a named \code{list} with 2 elements, each
being a vector of \code{numeric} data.
}

Optionally, \code{x} can have an attribute called \code{"meta"}, which must be a named list.
}
\examples{
# Known parameter values
d_cont(comparison_data$stature1)

# Average, standard deviation (SD), and sample size
d_cont(comparison_data$stature2)
d_cont(comparison_data$Vitamin.D)
d_cont(comparison_data$cholesterol)

# Average, standard error of the mean (SEM), and sample size
d_cont(comparison_data$serotonin)
}
\seealso{
Other es: 
\code{\link{d_bin}()},
\code{\link{d_np}()},
\code{\link{d_one}()},
\code{\link{d_ord}()},
\code{\link{d_pair}()}
}
\concept{es}
