% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_np.R
\name{d_np}
\alias{d_np}
\title{Dominance effect sizes using non-parametric estimation}
\usage{
d_np(x, meta = NULL)
}
\arguments{
\item{x}{data. See details.}

\item{meta}{either \code{NULL} or a named list.}
}
\value{
Estimates with verbose explanations.
}
\description{
Calculate dominance-based effect sizes using non-parametric estimation. Can
be used with continuous, binary, or ordinal data.
}
\details{
Calculate the following:
\itemize{
\item stochastic difference (aka Glass rank-biserial correlation);
\item adjusted stochastic difference (aka Goodman and Kruskal's gamma);
\item dominance (aka common language effect size);
\item measure of stochastic superiority (aka Vargha and Delaney's A);
\item adjusted dominance;
\item generalized odds ratio; and
\item win odds.
}

\code{x} can contain microdata or certain summary statistics.
\itemize{
\item If \code{x} contains microdata, it should be a named \code{list} with 2 elements, each
being a vector of data.
\item If \code{x} contains summary statistics, it should be a \code{data.frame} with 2 rows.
It can optionally contain a variable called \code{group}, indicating the name of the group.
In addition, it must contain one of the following:
\itemize{
\item \strong{Binary data.} The number of events (\code{k}) (also called "successes") and the
number of experiments or trials (\code{n}). Or
\item \strong{Ordinal data.} One variable for each possible outcome (2 or more outcomes).
The outcome variables \strong{must be ordered from best to worst}. The outcome
variables must contain the number of events for that outcome.
}
}

Optionally, \code{x} can have an attribute called \code{"meta"}, which must be a named list.
}
\examples{
# Binary data (number of events and sample size). Compare with `d_bin()`.
d_np(comparison_data$UTI)
d_np(comparison_data$DVT)
d_np(comparison_data$MI)

# Ordinal data (number of events for each outcome). Compare with `d_ord()`.
d_np(comparison_data$TB)
d_np(comparison_data$tonsils)
}
\seealso{
Other es: 
\code{\link{d_bin}()},
\code{\link{d_cont}()},
\code{\link{d_one}()},
\code{\link{d_ord}()},
\code{\link{d_pair}()}
}
\concept{es}
