% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_one.R
\name{d_one}
\alias{d_one}
\title{Dominance effect sizes for one-sample continuous variables}
\usage{
d_one(x, meta = NULL, .input = NULL)
}
\arguments{
\item{x}{data. See details.}

\item{meta}{either \code{NULL} or a named list.}

\item{.input}{(don't use)}
}
\value{
Estimates with verbose explanations.
}
\description{
Calculate dominance-based effect sizes for continuous (normally distributed)
variables in the one-sample or paired scenario.
}
\details{
Calculate the following:
\itemize{
\item dominance (aka common language effect size);
\item stochastic difference (aka Glass rank-biserial correlation);
\item odds, which, for continuous variables, is a special case of generalized odds ratio; and
\item standardized mean difference (aka Cohen's d).
}

For continuous variables, the following effect sizes simplify to one of the
above-mentioned effect sizes. Thus, they are not reported separately.
\itemize{
\item measure of stochastic superiority (aka Vargha and Delaney's A), which simplifies to dominance;
\item adjusted dominance, which simplifies to dominance; and
\item adjusted stochastic difference (aka Goodman and Kruskal's gamma), which simplifies to stochastic difference; and
\item win odds, which simplifies to odds.
}

\code{x} can contain either known parameter values, summary statistics, or microdata.
\itemize{
\item If \code{x} contains known parameter values or summary statistics, it should be
a \code{data.frame} with 2 rows. It can optionally contain a variable called \code{group},
indicating the name of the group. In addition, it must contain one of the following:
\itemize{
\item expected value (\code{mu}) and standard deviation (\code{sigma}), which are known parameters; or
\item sample mean (\code{xbar}), sample standard deviation (\code{sd}), and sample size (\code{n}); or
\item \code{xbar}, standard error of the mean (\code{sem}), and \code{n}.
}

\code{sd} is the "sample" standard deviation, with denominator \code{n - 1}. This is
what is returned by the \code{sd()} function.
}

One of the rows describes the population. The other row describes the constant.
For the row that describes the constant, either \code{mu} or \code{xbar} is equal to the
constant, while the other variables are set to \code{NA}.
\itemize{
\item If \code{x} contains microdata, it should be a named \code{list} with 2 elements, each
being a vector of \code{numeric} data. The element describing the constant must have
length \code{1}, while the element describing the population must have a length of
\verb{> 1}.
}

Optionally, \code{x} can have an attribute called \code{"meta"}, which must be a named list.
}
\examples{
d_one(comparison_data$stroke)

# Paired data. Compare with `d_pair()`.
df1 = rbind(comparison_data$IOP, data.frame(group = "Zero", xbar = 0, sd = NA, n = NA))
d_one(df1)
}
\seealso{
Other es: 
\code{\link{d_bin}()},
\code{\link{d_cont}()},
\code{\link{d_np}()},
\code{\link{d_ord}()},
\code{\link{d_pair}()}
}
\concept{es}
