% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_ord.R
\name{d_ord}
\alias{d_ord}
\title{Dominance effect sizes for ordinal variables}
\usage{
d_ord(x, meta = NULL)
}
\arguments{
\item{x}{data. See details.}

\item{meta}{either \code{NULL} or a named list.}
}
\value{
Estimates with verbose explanations.
}
\description{
Calculate dominance-based effect sizes for ordinal variables.
}
\details{
Calculate the following:
\itemize{
\item stochastic difference (aka Glass rank-biserial correlation);
\item adjusted stochastic difference (aka Goodman and Kruskal's gamma);
\item dominance (aka common language effect size);
\item measure of stochastic superiority (aka Vargha and Delaney's A);
\item adjusted dominance; and
\item generalized odds ratio; and
\item win odds.
}

\code{x} can contain either known parameter values or observed counts.

\code{x} should be a \code{data.frame} with 2 rows. It can optionally contain a variable called \code{group},
indicating the name of the group. In addition, it must contain one variable for each
possible outcome (2 or more outcomes). The outcome variables \strong{must be ordered from best to worst}.

The outcome variables can contain one of the following:
\itemize{
\item event probabilities, which are known parameters, and which must be between 0 and 1 and add up to 1; or
\item the number of events.
}

In the special case when there are exactly two outcome variables and they are called
\code{k} and \code{n}, this is taken to be binary data: \code{k} is the number of events;
\code{n} is the sample size -- it is the number of events plus the number of non-events.

Optionally, \code{x} can have an attribute called \code{"meta"}, which must be a named list.
}
\examples{
# Number of events for each outcome
d_ord(comparison_data$TB)
d_ord(comparison_data$tonsils)

# Binary data (number of events and sample size). Compare with `d_bin()`.
d_ord(comparison_data$UTI)

# Known parameter values (based on the tonsils example)
d_ord(data.frame(a1 = c(497 / 1326, 19 / 72), a2 = c(560 / 1326, 29 / 72)
, a3 = c(269 / 1326, 24 / 72)))
}
\seealso{
Other es: 
\code{\link{d_bin}()},
\code{\link{d_cont}()},
\code{\link{d_np}()},
\code{\link{d_one}()},
\code{\link{d_pair}()}
}
\concept{es}
