% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es4all-package.R
\docType{package}
\name{es4all-options}
\alias{es4all-options}
\title{Package options}
\description{
Typical users should see \code{\link[=set_opts]{set_opts()}} and \code{\link[=show_opts]{show_opts()}} to set and show certain options.
}
\details{
To view all available options, use \code{\link[=show_options]{show_options()}}.
}
\seealso{
Other options: 
\code{\link{set_opts}()},
\code{\link{show_options}()}
}
\author{
\strong{Maintainer}: [REDACTED] [REDACTED] \email{REDACTED@example.com} (\href{https://orcid.org/[REDACTED]}{ORCID})

}
\concept{options}
