% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es4all-package.R
\docType{package}
\name{es4all-package}
\alias{es4all}
\alias{es4all-package}
\title{es4all: Effect Sizes for All Data Types with CI's and P-Values}
\description{
Effect sizes for all variable types: continuous, binary, and ordinal. Clear explanations to convey the practical significance of differences to both statistical and non-statistical audiences. Text snippets that authors can directly incorporate into the Methods and Results sections of their research papers. Promote transparent and accessible communication of empirical findings while fostering evidence-based decision making across disciplines. Includes common language effect size, generalized odds ratio, Cliff's delta, Vargha and Delaney's A, Goodman and Kruskal's gamma, and other dominance-based effect sizes. Confidence intervals and p-values / hypothesis tests.
}
\author{
\strong{Maintainer}: [REDACTED] [REDACTED] \email{REDACTED@example.com} (\href{https://orcid.org/[REDACTED]}{ORCID})

}
\keyword{internal}
