% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_properties.R
\name{estimator_properties}
\alias{estimator_properties}
\title{Estimator properties of dominance-based effect sizes}
\usage{
estimator_properties(
  x,
  nsim,
  seed = 42,
  Sys.sleep_time = if (nsim > 100) 5 else 0
)
}
\arguments{
\item{x}{a 2-row \code{data.frame} with simulation parameters}

\item{nsim}{number of simulations}

\item{seed}{random number seed (ignored if \code{NA})}

\item{Sys.sleep_time}{occasional calls to \code{Sys.sleep()}}
}
\value{
Estimator properties.
}
\description{
Calculate the root mean squared error (RMSE), coverage, power, and
false positive rate (FPR) of dominance-based effect size estimators -- the
maximum likelihood estimator (MLE) and the non-parametric estimator (NPE).
Compares power and FPR with that of certain conventional tests.

\strong{Warning.} This function takes a very long time to run as it uses simulations.
}
\details{
\code{x} is a \code{data.frame} with 2 rows that specify how data is generated. It
has a variable called \code{n} (sample size). In addition, it has the following variables:
\itemize{
\item Continuous (normally distributed) data: \code{mu} and \code{sigma}.
\item Binary data: \code{p}.
}
}
\examples{
\donttest{
# **Warning.** This function takes a very long time to run as it uses simulations.
# In a real run, update the value of `nsim`.
# These examples have a theoretical power of 80\% and a theoretical FPR of 5\%.
nsim = 2 ## 5000
estimator_properties(data.frame(mu = c(105, 100), sigma = c(10, 10), n = c(64, 64)), nsim = nsim)
estimator_properties(data.frame(p = c(0.6, 0.4), n = c(97, 97)), nsim = nsim)
}
}
\concept{properties}
