% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.astra_table.R
\name{print.astra_table}
\alias{print.astra_table}
\alias{print.astra_list}
\title{Print astra tables}
\usage{
\method{print}{astra_table}(x, ...)

\method{print}{astra_list}(x, ...)
}
\arguments{
\item{x}{an object of class \code{astra_table} or \code{astra_list}.}

\item{...}{passed to helper functions.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print astra tables
}
\examples{
print( d_bin(comparison_data$UTI) )
}
\seealso{
Other print: 
\code{\link{set_opts}()}
}
\concept{print}
