% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_opts.R, R/show_opts.R
\name{set_opts}
\alias{set_opts}
\alias{show_opts}
\title{Set certain options}
\usage{
set_opts(
  reset = NULL,
  conf.level = NULL,
  es = NULL,
  output = NULL,
  file = NULL,
  .file_temp = NULL,
  raw = NULL
)

show_opts()
}
\arguments{
\item{reset}{reset all options to their default values?}

\item{conf.level}{confidence level of the interval.}

\item{es}{which effect sizes to print? \code{NA} (the default, prints all), or a
vector of short effect size names. See details.}

\item{output}{specify how the output is printed: \code{"auto"} (default); \code{"huxtable"}, \code{"gt"},
\code{"kableExtra"}, \code{"flextable"}; \code{"screen"}. For the following output types, please
also specify the \code{file} argument: \code{"Excel"}, \code{"Word"}, \code{"CSV"}.}

\item{file}{file name (see \code{output}).}

\item{.file_temp}{place \code{file} in a temporary folder?}

\item{raw}{if \code{TRUE}, \verb{d_*()} functions return the raw intermediate
\code{data.frame} instead of the formatted output. Default: \code{FALSE}.}
}
\value{
(Nothing.)
}
\description{
\code{set_opts()} sets certain package options. To view these options, use \code{show_opts()}.
For more advanced control and detailed customization, experienced  users can
also employ \code{\link[=options]{options()}} and \code{\link[=show_options]{show_options()}} (refer to \link{es4all-options}
for further information).
}
\details{
If you are not setting a particular option, leave it as \code{NULL}.

If \code{es = NA}, print all effect sizes. If \code{es} is a vector of
short effect size names, print just those effect sizes. Possible values are:
\verb{"SD", "ASD", "NNT", "D", "MSS", "AD", "GOR", "WO", "SMD"}.

\code{output} determines how the output is printed:
\itemize{
\item \code{"auto"} (default): automatically select the table-making package, depending on the
destination (such as screen, HTML, or PDF / LaTeX).
\item \code{"huxtable"}, \code{"gt"}, \code{"kableExtra"}, \code{"flextable"}: use this table-making package. Be sure
that this package is installed.
\item \code{"screen"}: print plain-text tables to the screen.
\item \code{"Excel"}: print to an Excel workbook. Please specify the name of an Excel
file using the \code{file} argument. Before using Excel printing, please be sure to install
the \code{openxlsx2} package.
\item \code{"Word"}: print to a Word document. Please specify the name of a Word
file using the \code{file} argument. Before using Word printing, please be sure to install these
packages: \code{flextable} and \code{officer}.
\item \code{"CSV"}: print to a comma-separated values (CSV) file. Please specify the name of a
CSV file using the \code{file} argument.
}
}
\examples{
set_opts(es = NA)
}
\seealso{
Other options: 
\code{\link{es4all-options}},
\code{\link{show_options}()}

Other print: 
\code{\link{print.astra_table}()}
}
\concept{options}
\concept{print}
