% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_options.R
\name{show_options}
\alias{show_options}
\title{Show package options}
\usage{
show_options(sw = c("es4all", "astra"))
}
\arguments{
\item{sw}{starting characters}
}
\value{
List of options and their values.
}
\description{
See \link{es4all-options} for a discussion of some of the options.
}
\examples{
show_options()
}
\seealso{
Other options: 
\code{\link{es4all-options}},
\code{\link{set_opts}()}
}
\concept{options}
